fix: moves dbUtils.js inside src
This makes it easier to identify this script as part of the source code. Also, it's easier to make the src folder as a volume for docker compose for realtime update of the contents while the container is running.
